test: fix Issue #16878 test case structure and logic
